awesome: refactored
